First step towards highlight search query on detail page
