+++ src/stores/search.ts
@@ -1,10 +1,17 @@
 import { create, StateCreator } from "zustand";
-import { SearchResult } from "../model/Search";
+import { SearchQuery, SearchResult } from "../model/Search";
 
 export type SearchResultsSlice = {
   globalSearchResults: SearchResult | undefined;
   setGlobalSearchResults: (
     newSearchResults: SearchResultsSlice["globalSearchResults"],
+  ) => void;
+};
+
+export type SearchQuerySlice = {
+  globalSearchQuery: SearchQuery | undefined;
+  setGlobalSearchQuery: (
+    newSearchQuery: SearchQuerySlice["globalSearchQuery"],
   ) => void;
 };
 
@@ -19,6 +26,20 @@ const createSearchResultsSlice: StateCreator<
     set(() => ({ globalSearchResults: newSearchResults })),
 });
 
-export const useSearchStore = create<SearchResultsSlice>()((...a) => ({
-  ...createSearchResultsSlice(...a),
-}));
+const createSearchQuerySlice: StateCreator<
+  SearchQuerySlice,
+  [],
+  [],
+  SearchQuerySlice
+> = (set) => ({
+  globalSearchQuery: undefined,
+  setGlobalSearchQuery: (newSearchQuery) =>
+    set(() => ({ globalSearchQuery: newSearchQuery })),
+});
+
+export const useSearchStore = create<SearchResultsSlice & SearchQuerySlice>()(
+  (...a) => ({
+    ...createSearchResultsSlice(...a),
+    ...createSearchQuerySlice(...a),
+  }),
+);
